cleanup: build with -Wconversion (#4907)
Enable `-Wconversion -Wno-sign-conversion` on the CMake builds. We
discussed this as maybe introducing too many casts, but also maybe
helping us catch stuff.

+++ google/cloud/spanner/benchmarks/multiple_rows_cpu_benchmark.cc
@@ -1688,8 +1688,8 @@ void SimpleTimer::Stop() {
   cpu_time_ = utime + stime;
   double cpu_fraction = 0;
   if (elapsed_time_.count() != 0) {
-    cpu_fraction =
-        (cpu_time_).count() / static_cast<double>(elapsed_time_.count());
+    cpu_fraction = static_cast<double>(cpu_time_.count()) /
+                   static_cast<double>(elapsed_time_.count());
   }
   now.ru_minflt -= start_usage_.ru_minflt;
   now.ru_majflt -= start_usage_.ru_majflt;
